+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.0.0 (2021-06-21)
+# 0.0.0 (2021-06-22)
 
 ### Bug Fixes
 
@@ -56,8 +56,10 @@
 - **algorithm/classic:** add minDistance ([f441103](https://github.com/air-supply94/js-algorithm/commit/f44110326d1e4c037a81176cc8041be436ff5b68))
 - **algorithm/classic:** add rubberHouse ([bf41594](https://github.com/air-supply94/js-algorithm/commit/bf41594753015b1d45cb67ed4ef42600805f5f8e))
 - **algorithm/math:** add findErrorNums ([edee6fb](https://github.com/air-supply94/js-algorithm/commit/edee6fb2edebd8696f61c8ca66e0e8de77d46a3d))
+- **algorithm/math:** add game ([20dfd5c](https://github.com/air-supply94/js-algorithm/commit/20dfd5c1c6c15e47220442bcba3d0393bfcbadc9))
 - **algorithm/math:** add getRandom ([da6ffd9](https://github.com/air-supply94/js-algorithm/commit/da6ffd976e772de2d5b6b6903ec597a752391bf1))
 - **algorithm/math:** add missingNumber ([168e330](https://github.com/air-supply94/js-algorithm/commit/168e33009a3ad21ba4a15c56c65367588e6c03ad))
+- **algorithm/math:** add subarraySum ([8f23792](https://github.com/air-supply94/js-algorithm/commit/8f23792c76acb6f8f2d60d9e6ed24a117f14e4ce))
 - **algorithm/math:** add trailingZeroes ([cefab4c](https://github.com/air-supply94/js-algorithm/commit/cefab4c03594616e8c8ff8ed447a4bdd1febff3d))
 - **algorithm/math:** add WSRARes ([d6589c4](https://github.com/air-supply94/js-algorithm/commit/d6589c43f9c6dd469b0a642b4df9f43a136de3b8))
 - **algorithm/slidingWindow:** add checkInclusion ([e952fc8](https://github.com/air-supply94/js-algorithm/commit/e952fc852e3392f5a1e92fdbb6cd1fa02d20d8d7))